pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "acs-sdk"
version = "0.1.0"
description = "Python SDK for Apache CloudStack API"
authors = [
    {name = "Vimal Saini",email = "vimal.saini25@gmail.com"}
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
dependencies = [
    "httpx (>=0.28.1,<0.29.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "click (>=8.3.1,<9.0.0)",
    "opentelemetry-sdk (>=1.40.0,<2.0.0)"
]

[project.scripts]
acs = "acs_sdk.cli:console"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "ruff (>=0.15.7,<0.16.0)",
    "mypy (>=1.19.1,<2.0.0)"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-ra -q --cov=acs_sdk --cov-report=term-missing"


[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
strict = true